+++ otto/models.py
@@ -1,26 +1,6 @@
 from pydantic import BaseModel, AnyUrl
 from typing import List, Dict, Optional, Union
-from otto.as_form import as_form
 
-@as_form
-class VideoForm(BaseModel):
-    project: str = ''
-    name: str = ''
-    logo: str = '' #AnyUrl
-    address: str = ''
-    phone: str = ''
-    hours: str = ''
-    website: str = '' #AnyUrl = 'talahairstudio.com'
-    initial: str = ''
-    bullets: str = ''
-    media: str = ''
-    audio: str = ''
-    call: str = ''
-    closing: str = ''
-    fontcolor: str = '#FFFFFF'
-    themecolor: str = '#CC5500'
-    font: str = 'Segoe_UI_Bold'
-    duration: float = 5
 
 class Clip(BaseModel):
     duration: float = 5
