+++ otto/render.py
@@ -8,41 +8,7 @@ from otto.kburns import kburns
 from otto.models import Edl
 from typing import List
 
-def renderEdl(edl: Edl, media, audio=None, filename='render.mp4', moviesize=(1920,1080), logger='bar'):
-    clips = []
-    duration = 0
-    media = [ download(m) for m in media ]
-    for clip in edl.clips:
-        print('making clip', clip, type(clip))
-        duration += clip.duration
-        if clip.type == 'template':
-            tmp = getattr(templates, clip.name)
-            print('making template', tmp )
-            clips.append(
-                tmp(**clip.data, duration=clip.duration, clipsize=moviesize)
-            )
-        elif clip.type == 'video':
-            clips.append(
-                VideoFileClip(clip.name)
-                .subclip(clip.inpoint)
-                .set_duration(clip.duration)
-            )
-    print('made clips', clips)
-    video = concatenate_videoclips(clips).resize(moviesize)
-    print('made video', video, duration)
-    kburns(media[:int(1 + duration//5)], moviesize=moviesize, filename=f'{filename}_kbout.mp4')
-    slides = (VideoFileClip(f'{filename}_kbout.mp4')
-            .set_duration(duration)
-            .crossfadein(1)
-            .crossfadeout(1)
-    )
-    video = CompositeVideoClip([slides, video])
-    if audio:
-        video = video.set_audio(AudioFileClip(audio))
-    video = video.set_duration(duration).crossfadeout(1).audio_fadeout(1)
-    video.write_videofile(filename, fps=30, logger=logger, threads=8)
-
-def generateEdl(edl: Edl, moviesize=(1920,1080), audio=None, duration=None, inpoint=None, **kwargs):
+def generateEdl(edl: Edl, moviesize=(1920,1080), audio=None, **kwargs):
     """Generates a moviepy CompositeVideoClip from an Edl"""
     clips = []
     for clip in edl.clips:
@@ -107,7 +73,3 @@ def renderMultitrack(edl: Edl,
         audio_fps=48000, audio_codec='aac', audio_bitrate=audio_bitrate,
         codec=codec, bitrate=bitrate,
         ffmpeg_params=ffmpeg_params)
-
-def renderForm(form, filename='render.mp4', logger='bar'):
-    v = Otto(form)
-    v.render().write_videofile(filename=filename, fps=fps, logger=logger)
